backend/requirements.txt
fastapi
uvicorn
python-docx
python-multipart
requests
openai-whisper
pydub
python-dotenv
